Slight tweak
isExpanded should be 0 instead of true:

The key change here is setting isExpanded to 0 instead of true. In Vuetify's v-expansion-panels, when using v-model:
null means all panels are collapsed
A number (like 0) means that panel at that index is expanded
An array can be used for multiple panels
This change will ensure the panel is expanded by default when the component is mounted. The rest of your code can remain the same since you're already handling the icon toggle based on this value.